Add output utilities and refactor tests
